Cleaned up the TODOs and updated Login response and dashboard response model according to backend data format

+++ lib/screens/delivery_partener/model/delivery_order_model.dart
@@ -105,7 +105,7 @@ class DeliveryStats {
   final int totalDeliveries;
   final double totalEarnings;
   final double rating;
-  final int onlineHours;
+  final double onlineHours;
   final double acceptanceRate;
 
   DeliveryStats({
